Move Cropper to Statamic\CP\Assets\CropProcessor
Statamic\Imaging is the public imaging surface addon devs reach for. This
helper is CP-only and does more than crop (format conversion, quality,
transparency flattening), so the CP\Assets namespace and "Processor" name
signal it's internal. Marked @internal.
