+++ src/Http/View/Composers/JavascriptComposer.php
@@ -7,6 +7,7 @@ use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 use Statamic\Assets\CropAspectRatios;
 use Statamic\CommandPalette\Category;
+use Statamic\CP\Assets\CropProcessor;
 use Statamic\CP\Color;
 use Statamic\Facades\CommandPalette;
 use Statamic\Facades\CP\Toast;
@@ -16,7 +17,6 @@ use Statamic\Facades\Site;
 use Statamic\Facades\User;
 use Statamic\Fieldtypes\Sets;
 use Statamic\Icons\IconSet;
-use Statamic\Imaging\Cropper;
 use Statamic\Statamic;
 use Statamic\Support\Str;
 use voku\helper\ASCII;
@@ -67,7 +67,7 @@ class JavascriptComposer
             'googleDocsViewer' => config('statamic.assets.google_docs_viewer'),
             'focalPointEditorEnabled' => config('statamic.assets.focal_point_editor'),
             'cropAspectRatios' => CropAspectRatios::all(),
-            'cropQuality' => Cropper::defaultQuality(),
+            'cropQuality' => CropProcessor::defaultQuality(),
             'elevatedSessionsEnabled' => config('statamic.users.elevated_sessions_enabled'),
             'user' => $this->user($user),
             'defaultPreferences' => Preference::default()->all(),
